demo/poc-v1/examples/python-caster/requirements.txt
grpcio==1.72.0
grpcio-tools==1.72.0
protobuf==5.29.4
